Split runtime cache domains
