+++ policyengine_api/services/household_calculation_service.py
@@ -17,7 +17,7 @@ from policyengine_api.data.v1_models import (
 )
 from policyengine_api.runtime_cache.dependencies import get_runtime_cache_context
 from policyengine_api.runtime_cache.core import record_cache_event
-from policyengine_api.runtime_cache.repositories import (
+from policyengine_api.runtime_cache.household_traces import (
     HouseholdTraceCache,
     HouseholdTraceIdentity,
     HouseholdTraceValue,
