+++ tests/unit/routes/test_household_and_user_policy_orm_routes.py
@@ -13,7 +13,7 @@ from policyengine_api.data.v1_models import (
 )
 from policyengine_api.runtime_cache.core import CacheNamespace
 from policyengine_api.runtime_cache.fake import InMemoryCacheBackend
-from policyengine_api.runtime_cache.repositories import (
+from policyengine_api.runtime_cache.household_traces import (
     HouseholdTraceCache,
     HouseholdTraceIdentity,
     HouseholdTraceValue,
