+++ tests/unit/services/test_household_calculation_service.py
@@ -10,7 +10,7 @@ from policyengine_api.data.v1_models import (
 )
 from policyengine_api.runtime_cache.core import CacheNamespace
 from policyengine_api.runtime_cache.fake import InMemoryCacheBackend
-from policyengine_api.runtime_cache.repositories import (
+from policyengine_api.runtime_cache.household_traces import (
     HouseholdTraceCache,
     HouseholdTraceIdentity,
     HouseholdTraceValue,
